Remove xs_float

+++ sys_stats.h
@@ -5,8 +5,8 @@
 
 typedef struct {
   double moments[3];
-  uint64_t min;
-  uint64_t max;
+  double min;
+  double max;
 } Stats_t;
 
 void sys_statsInit(Stats_t *accum);
@@ -17,4 +17,4 @@ double sys_statsVariance(Stats_t *accum);
 double sys_statsMin(Stats_t *accum);
 double sys_statsMax(Stats_t *accum);
 
-void sys_statsAddSample(Stats_t *accum, uint64_t sample);
+void sys_statsAddSample(Stats_t *accum, double sample);
